requirements_im.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
firebase_admin==7.0.0
frozenlist==1.7.0
gmalg==1.0.6
gmssl==3.2.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
lazy_loader==0.4
msgpack==1.1.1
multidict==6.6.3
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
PyJWT==2.10.1
PyPDF2==3.0.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.0
simple-websocket==1.1.0
sniffio==1.3.1
tifffile==2025.6.11
urllib3==2.5.0
websockets==12.0
wsproto==1.2.0
yarl==1.20.1
